Default `To` address if X-Original-To is not given
GMAIL by default does not send the 'X-Original-To' header in an email for the initial address. As a result, the mail to the server won't process since there is no X-Original-To address to get the recipient email from. Therefore, I've added `msg.get('To')` as a fallback method to get the recipient address.